resolve comment nic
